# simplify validateMimeValue regex